scripts/requirements_vertex_ai.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.38.0
